Add public_keys attribute (#562)
This adds a `public_keys` attribute to `ec2_metadata`. Like in Boto 2,
it's a dictionary keyed by the human-assigned SSH key name.

Whereas Boto 2 always retrieved the `/openssh-key` field for every key
and represented the key as a string (see
https://github.com/boto/boto/blob/70c65b4f67af41ccfd40d21e49880be568997ba6/boto/utils.py#L261-L265),
this introduces a `PublicKey` object with an `openssh_key` attribute to
model the single-field object. The docs at
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html
document just a `public-keys/0/openssh-key` endpoint. The API structure
is designed to allow for multiple keys and multiple formats, and Boto 2
already supported multiple keys, so I wrote this to support both
multiple keys and multiple formats even though only `openssh-key`
currently exists.

Instead of a `None` for the case where the list of keys is not available
at all, I'm returning an empty dict `{}`.

I'm making an assumption that what I'm calling the "index" of a key (`0`
in all of Amazon's examples) is constrained to be an integer. Right now
I think Amazon only ever produces 0.

I've written tests for this that pass, and I've manually tested on an
AWS instance and was able to retrieve the key. I noticed that AWS serves
the key data with a trailing newline, which I'm not making any attempt
to remove.

This should fix #561.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Adam Johnson <me@adamj.eu>